Fix: 충돌오류해결 #11

+++ admin/src/main/java/com/sh/admin/item/model/dao/ItemMapper.java
@@ -13,7 +13,7 @@ import java.util.List;
 public interface ItemMapper {
     List<ItemDto> findAll(); //  전체조회
 
-    ItemDto findById(int itemId); // 아이디로 한건조회
+    ItemDto findById(long itemId); // 아이디로 한건조회
 
 //    int updateSaleStatus(int itemId);
     /* 희윤 시작 */
